% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_model.R
\name{gt_model}
\alias{gt_model}
\title{Generate a gt table from model results.}
\usage{
gt_model(
  models,
  digits = 3,
  sig_thresh = 0.05,
  var_labels = NULL,
  summary_stats = NULL
)
}
\arguments{
\item{models}{a \code{list} of models, of either the \code{lm} or \code{glm} command.}

\item{digits}{a numeric value indicating the number of decimals to round results to}

\item{sig_thresh}{a numeric value indicating the threshold for statistical significance
for the asterisk. If NULL, asterisks will not be printed.}

\item{var_labels}{a named character vector indicating labels for the rows. Names should
be the actual row names and values should be the labels desired.}

\item{summary_stats}{a character vector indicating desired summary statistics. See below
for a list of available options.}
}
\value{
\code{gt_model} returns a \code{gt_tbl} object that can be further processed using
various commands from the \code{gt} package.
}
\description{
A description
}
\details{
This function can be used to create a \code{gt_tbl} table of model (i.e. regression) results
using a format common in many scientific fields. Multiple models can be included in the table
by feeding in a list of model objects. The most common intended use of this function is within
Currently the following models are verified to work:

Because the returned object is a \code{gt_tbl}, it can be further refined to the user's
tastes by piping it into subsequent \code{gt} commands.

The returned table uses a label row for the variable label and summary statistics labels. All
Users can rename the models with a \code{cols_label} command. Variables and summary statistics
can be renamed by the \code{var_labels} argument which takes a named character vector that
provides the correspondence between original and new variable names (see examples below).
}
